Moved comments navigation hook
no issue

The comments navigation hook is content-specific, so it now follows the feature hooks directory pattern used by the admin app instead of living beside components.
